Update run.cfc to use webroot when resolving testbox location in ensureTestBox
Currently, if your testbox folder is not in the current folder where your server.json lives, ie, a subfolder like /test-harness or /app the ensureTestBox() function fails, and then installs testbox into the testbox-cli folder.

In CI, this testbox-cli folder never has a testbox in it, because its built from an image, so this means every time we run tests, it installs testbox over and over and over again.

This fix users the server info to pass the webroot of the current server to the resolvePath() command, to ensure we look for testbox in the webroot. 

Improvement: Maybe we look in the current folder, and in the webroot?